WIP: build(deps): Use latest github.com/containers/ocicrypt for pkcs11 support
This patch adds support for encrypting and decrypting images with
pkcs11 devices.

We need to add a configuration file from which pkcs11 parameters are read.
A config file may look like this:

pkcs11:
  module-directories:
  - /usr/lib64/pkcs11/
  allowed-module-paths:
  - /usr/lib64/pkcs11/libsofthsm2.so

The module-directories enumerates directories in which to look for pkcs11
modules. The allowed-module-paths restricts the pkcs11 modules that can
be used.

The config file is searched for in this order:
- ${OCICRYPT_CONFIG}="internal": use an internal allow-all policy
- ${OCICRYPT_CONFIG}
- ${XDG_CONFIG_HOME}/ocicrypt.conf
- ${HOME}/.config/ocicrypt.conf
- /etc/ocicrypt.conf

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>